Use calibration information in ComputeCurvatures and friends
A "FIXME" with this commit is that it currently assumes that
pixelWidth will have at least as small a value as pixelHeight
and pixelDepth.  I will fix in a later commit this so that the
default sigmas are based on the smallest of the three rather
than pixelWidth.

Other changes included in this commit:
  - Display progress properly in ComputeCurvatures_
  - More whitespace cleanup

+++ tracing/SearchThread.java
@@ -151,9 +151,9 @@ public abstract class SearchThread extends Thread {
 
 	ImagePlus imagePlus;
 
-	double x_spacing;
-	double y_spacing;
-	double z_spacing;
+	float x_spacing;
+	float y_spacing;
+	float z_spacing;
 
 	String spacing_units;
 
@@ -319,9 +319,9 @@ public abstract class SearchThread extends Thread {
 
 		Calibration calibration = imagePlus.getCalibration();
 
-		x_spacing = calibration.pixelWidth;
-		y_spacing = calibration.pixelHeight;
-		z_spacing = calibration.pixelDepth;
+		x_spacing = (float)calibration.pixelWidth;
+		y_spacing = (float)calibration.pixelHeight;
+		z_spacing = (float)calibration.pixelDepth;
 		spacing_units = calibration.getUnit();
 
 		if( (x_spacing == 0.0) ||
